test: Split the collect and transform tests into separate files
These are the last of the strategies whose tests were all grouped together
in the test_streaming.py file. Now each strategy has it's own set of tests.

None of the actual tests have changed here. Just moving them into the right files.
